% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterScore.R
\name{get_spatial_corr}
\alias{get_spatial_corr}
\title{Calculate the distance between histology annotation}
\usage{
get_spatial_corr(cellranger_out, annotation_table)
}
\arguments{
\item{cellranger_out}{10x spaceranger count folder}

\item{annotation_table}{table contains pathology_annotation, cluster, spot columns}
}
\value{
distance matrix
}
\description{
Calculate the distance between histology annotation
}
