% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterScore.R
\name{plot_c_involved}
\alias{plot_c_involved}
\title{Get the plot c involved path to the Python script file in the package, and call it}
\usage{
plot_c_involved(
  spatial_info_table,
  color_table,
  exp_table,
  highlight_PAs_fill = NULL,
  highlight_PAs_border = NULL,
  min_frac = 0.1,
  min_n = 10,
  output_folder = ".",
  output_file = "space_ha_c_involved.pdf"
)
}
\arguments{
\item{spatial_info_table}{cluster table path}

\item{color_table}{color table path}

\item{exp_table}{Path to a csv expression matrix file}

\item{highlight_PAs_fill}{clusters to highlight in fill color}

\item{highlight_PAs_border}{clusters to highlight in border color}

\item{min_frac}{minimum fraction}

\item{min_n}{minimum number}

\item{output_folder}{output folder path}

\item{output_file}{output file}
}
\description{
Get the plot c involved path to the Python script file in the package, and call it
}
