% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterScore.R
\name{plot_ha_cluster}
\alias{plot_ha_cluster}
\title{Get the plot ha cluster path to the Python script file in the package, and call it}
\usage{
plot_ha_cluster(
  spatial_info_table,
  color_table,
  highlight_PAs_fill = NULL,
  highlight_PAs_border = NULL,
  output_folder = ".",
  output_file = "space_ha_c.pdf"
)
}
\arguments{
\item{spatial_info_table}{cluster table path}

\item{color_table}{color table path}

\item{highlight_PAs_fill}{clusters to highlight in fill color}

\item{highlight_PAs_border}{clusters to highlight in border color}

\item{output_folder}{output folder path}

\item{output_file}{output file}
}
\description{
Get the plot ha cluster path to the Python script file in the package, and call it
}
