% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterScore.R
\name{SpatialClusterScore}
\alias{SpatialClusterScore}
\title{Calculate the cluster score for spatial transcriptomics data}
\usage{
SpatialClusterScore(spatial_info_table, D, alpha = 1, beta = 0)
}
\arguments{
\item{spatial_info_table}{Path to a tab delimited text file with column: spot, x_array, y_array,
pathology_annotation, cluster, pathology_annotation_type or, a tibble object
that contains those columns}

\item{D}{distance}

\item{alpha}{weight paramter 1}

\item{beta}{weight paramter 2}
}
\value{
a score in range [0,1]
}
\description{
Calculate the cluster score for spatial transcriptomics data
}
