% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterScore.R
\name{SpatialRogueScore}
\alias{SpatialRogueScore}
\title{Calculate the rogue score for spatial transcriptomics data}
\usage{
SpatialRogueScore(
  exp_table,
  spatial_info_table,
  min.cells = 10,
  min.genes = 10
)
}
\arguments{
\item{exp_table}{Path to a csv expression matrix file}

\item{spatial_info_table}{Path to a tab delimited text file with column: spot, x_array, y_array,
pathology_annotation, cluster, pathology_annotation_type or, a tibble object
that contains those columns}

\item{min.cells}{minimum cell (spot) number to calculate ROGUE, default 10}

\item{min.genes}{minimum gene number to calculate ROGUE, default 10}
}
\value{
a score in range [0,1]
}
\description{
Calculate the rogue score for spatial transcriptomics data
}
